refactor: use lazyquery for heart rate bpm breakdown
